reddit-queue/requirements.txt
discord.py
praw
